requirements.txt
requests==2.25.1
pyjwt==2.4.0
vantage6-client==3.7.3
pandas==1.3.5
scikit-learn==1.0.2
flake8==5.0.4